Chapter 1: Echo command line args update.
Changing the primitive for loop away from incrementing an integer
counter. The new for loop uses a range over the slice os.Args[].
The slice index is read into Go's bottom type, `_`, to be discarded
and the value is read into the variable `arg`. Accessing the range
of elements in the slice is done by skipping the zeroth element
and implicitly accessing all elements to the end of the slice by
the notation [1:]